Add reusable functions script file
