+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {
+      "lib": ["es5", "es6", "dom"],
+      "outDir": "./dist/",
+      "sourceMap": true,
+      "noImplicitAny": true,
+      "module": "commonjs",
+      "target": "es6",
+      "moduleResolution": "node",
+      "jsx": "react"
+    },
+    "include": ["./scripts"]
+}
+
